lptask0/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-crispy-forms==1.12.0
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.1
